requirements.txt
allure-pytest==2.8.18
pytest==6.2.5
requests==2.24.0
selenium==3.141.0
webdriver-manager==3.2.2